Extract local server from gulpfile to server/index.js
